requirements.txt
# Core Deep Learning Framework
tensorflow==2.13.0
keras==2.13.1

# Data Processing and Analysis
numpy==1.24.3
pandas==2.0.3
scikit-learn==1.3.0

# Image Processing
Pillow==10.0.0
opencv-python==4.8.0.76

# Visualization
matplotlib==3.7.2
seaborn==0.12.2

# Utilities
tqdm==4.65.0
python-dateutil==2.8.2

# Optional: GPU Support (uncomment if using NVIDIA GPU)
# tensorflow-gpu==2.13.0

# Development and Testing (optional)
# pytest==7.4.0
# jupyter==1.0.0
# ipykernel==6.24.0
